bet model validations

+++ alpha/backend-relay/migrations/20171102154214-create-bets.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       amount: {
-        type: Sequelize.FLOAT
+        type: Sequelize.INTEGER
       },
       edge: {
         type: Sequelize.FLOAT
